+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  - The "flare" spell now only spawns a firework effect
  - Allow two spells to use the same material icon.
  - Add a parameter to push/pull to specify how many entities to effect. Pull now does only 1 again (for Yoinking)
+ - Add "camera" spell, which takes a portrait of the target player.
 
 ## 2.6.2